Use the internal _mint() method to provide a custom supply

+++ irc2-token/src/test/java/com/iconloop/score/example/IRC2BasicTest.java
@@ -42,7 +42,7 @@ class IRC2BasicTest extends TestBase {
     @BeforeAll
     public static void setup() throws Exception {
         tokenScore = sm.deploy(owner, IRC2BasicToken.class,
-                name, symbol, BigInteger.valueOf(decimals), initialSupply);
+                name, symbol, decimals, initialSupply);
         owner.addBalance(symbol, totalSupply);
     }
